Add release version for building
Release version links with the currently added static versions of the
needed libraries. Currently, only glfw, freetype, and png need to be
statically linked.

Linux versions of the libraries aren't available at the moment.
Hopefully in the future I'll add a copy of these libraries to the
source, and then build them when I need static versions so it will match
the version it is built on, but not right now.